requirements.txt
beautifulsoup4
django-background-tasks
requests